third_party/blink/web_tests/platform/generic/fast/js/dfg-logical-not-final-object-or-other-expected.txt

Tests that logical not of an object where it is predicted either final object or other (i.e. null or undefined) performs correctly when document.all is present.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS foo(null) is false
PASS foo(o) is true
PASS successfullyParsed is true

TEST COMPLETE
